# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Course Generation Agent",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/threads/{thread_id}/files": {
      "post": {
        "summary": "Upload File",
        "operationId": "upload_file_api_v1_threads__thread_id__files_post",
        "parameters": [
          {
            "name": "thread_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Thread Id"
            }
          },
          {
            "name": "category",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/UploadCategory",
              "default": "context"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_file_api_v1_threads__thread_id__files_post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiEnvelope_UploadFileResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "List Files",
        "operationId": "list_files_api_v1_threads__thread_id__files_get",
        "parameters": [
          {
            "name": "thread_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Thread Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiEnvelope_ThreadFilesResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/threads/{thread_id}/artifacts/latest": {
      "get": {
        "summary": "Latest Artifact",
        "operationId": "latest_artifact_api_v1_threads__thread_id__artifacts_latest_get",
        "parameters": [
          {
            "name": "thread_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Thread Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiEnvelope_ArtifactResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "summary": "Update Latest Artifact",
        "operationId": "update_latest_artifact_api_v1_threads__thread_id__artifacts_latest_patch",
        "parameters": [
          {
            "name": "thread_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Thread Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ArtifactUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiEnvelope_ArtifactResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/threads/{thread_id}/artifacts/{version}": {
      "get": {
        "summary": "Get Artifact Version",
        "operationId": "get_artifact_version_api_v1_threads__thread_id__artifacts__version__get",
        "parameters": [
          {
            "name": "thread_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Thread Id"
            }
          },
          {
            "name": "version",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Version"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiEnvelope_ArtifactResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/threads/{thread_id}/artifacts/{version}/diff/{prev_version}": {
      "get": {
        "summary": "Artifact Diff",
        "operationId": "artifact_diff_api_v1_threads__thread_id__artifacts__version__diff__prev_version__get",
        "parameters": [
          {
            "name": "thread_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Thread Id"
            }
          },
          {
            "name": "version",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Version"
            }
          },
          {
            "name": "prev_version",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Prev Version"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiEnvelope_ArtifactDiffResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/threads/{thread_id}/review-batches/{batch_id}": {
      "get": {
        "summary": "Get Review Batch",
        "operationId": "get_review_batch_api_v1_threads__thread_id__review_batches__batch_id__get",
        "parameters": [
          {
            "name": "thread_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Thread Id"
            }
          },
          {
            "name": "batch_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Batch Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiEnvelope_ReviewBatchResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/threads/{thread_id}/review-batches/{batch_id}/submit": {
      "post": {
        "summary": "Submit Review",
        "operationId": "submit_review_api_v1_threads__thread_id__review_batches__batch_id__submit_post",
        "parameters": [
          {
            "name": "thread_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Thread Id"
            }
          },
          {
            "name": "batch_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Batch Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewSubmitRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiEnvelope_ReviewSubmitResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/threads/{thread_id}/decision-records": {
      "get": {
        "summary": "List Thread Decision Records",
        "operationId": "list_thread_decision_records_api_v1_threads__thread_id__decision_records_get",
        "parameters": [
          {
            "name": "thread_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Thread Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiEnvelope_DecisionRecordsResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ApiEnvelope_ArtifactDiffResponse_": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "default": true
          },
          "request_id": {
            "type": "string",
            "title": "Request Id"
          },
          "thread_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Thread Id"
          },
          "data": {
            "$ref": "#/components/schemas/ArtifactDiffResponse"
          },
          "error": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "meta": {
            "additionalProperties": true,
            "type": "object",
            "title": "Meta"
          }
        },
        "type": "object",
        "required": [
          "request_id",
          "data"
        ],
        "title": "ApiEnvelope[ArtifactDiffResponse]"
      },
      "ApiEnvelope_ArtifactResponse_": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "default": true
          },
          "request_id": {
            "type": "string",
            "title": "Request Id"
          },
          "thread_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Thread Id"
          },
          "data": {
            "$ref": "#/components/schemas/ArtifactResponse"
          },
          "error": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "meta": {
            "additionalProperties": true,
            "type": "object",
            "title": "Meta"
          }
        },
        "type": "object",
        "required": [
          "request_id",
          "data"
        ],
        "title": "ApiEnvelope[ArtifactResponse]"
      },
...
      "ApiEnvelope_DecisionRecordsResponse_": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "default": true
          },
          "request_id": {
            "type": "string",
            "title": "Request Id"
          },
          "thread_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Thread Id"
          },
          "data": {
            "$ref": "#/components/schemas/DecisionRecordsResponse"
          },
          "error": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "meta": {
            "additionalProperties": true,
            "type": "object",
            "title": "Meta"
          }
        },
        "type": "object",
        "required": [
          "request_id",
          "data"
        ],
        "title": "ApiEnvelope[DecisionRecordsResponse]"
      },
...
      "ApiEnvelope_ReviewBatchResponse_": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "default": true
          },
          "request_id": {
            "type": "string",
            "title": "Request Id"
          },
          "thread_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Thread Id"
          },
          "data": {
            "$ref": "#/components/schemas/ReviewBatchResponse"
          },
          "error": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "meta": {
            "additionalProperties": true,
            "type": "object",
            "title": "Meta"
          }
        },
        "type": "object",
        "required": [
          "request_id",
          "data"
        ],
        "title": "ApiEnvelope[ReviewBatchResponse]"
      },
      "ApiEnvelope_ReviewSubmitResponse_": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "default": true
          },
          "request_id": {
            "type": "string",
            "title": "Request Id"
          },
          "thread_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Thread Id"
          },
          "data": {
            "$ref": "#/components/schemas/ReviewSubmitResponse"
          },
          "error": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "meta": {
            "additionalProperties": true,
            "type": "object",
            "title": "Meta"
          }
        },
        "type": "object",
        "required": [
          "request_id",
          "data"
        ],
        "title": "ApiEnvelope[ReviewSubmitResponse]"
      },
...
      "ApiEnvelope_ThreadFilesResponse_": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "default": true
          },
          "request_id": {
            "type": "string",
            "title": "Request Id"
          },
          "thread_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Thread Id"
          },
          "data": {
            "$ref": "#/components/schemas/ThreadFilesResponse"
          },
          "error": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "meta": {
            "additionalProperties": true,
            "type": "object",
            "title": "Meta"
          }
        },
        "type": "object",
        "required": [
          "request_id",
          "data"
        ],
        "title": "ApiEnvelope[ThreadFilesResponse]"
      },
...
      "ApiEnvelope_UploadFileResponse_": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "default": true
          },
          "request_id": {
            "type": "string",
            "title": "Request Id"
          },
          "thread_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Thread Id"
          },
          "data": {
            "$ref": "#/components/schemas/UploadFileResponse"
          },
          "error": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "meta": {
            "additionalProperties": true,
            "type": "object",
            "title": "Meta"
          }
        },
        "type": "object",
        "required": [
          "request_id",
          "data"
        ],
        "title": "ApiEnvelope[UploadFileResponse]"
      },
      "ArtifactDiffResponse": {
        "properties": {
          "diff": {
            "type": "string",
            "title": "Diff"
          },
          "version": {
            "type": "integer",
            "title": "Version"
          },
          "prev_version": {
            "type": "integer",
            "title": "Prev Version"
          }
        },
        "type": "object",
        "required": [
          "diff",
          "version",
          "prev_version"
        ],
        "title": "ArtifactDiffResponse"
      },
      "ArtifactResponse": {
        "properties": {
          "artifact": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DraftArtifact"
              },
              {
                "$ref": "#/components/schemas/ArtifactVersionDetail"
              },
              {
                "type": "null"
              }
            ],
            "title": "Artifact"
          }
        },
        "type": "object",
        "required": [
          "artifact"
        ],
        "title": "ArtifactResponse"
      },
      "ArtifactUpdateRequest": {
        "properties": {
          "markdown": {
            "type": "string",
            "title": "Markdown"
          }
        },
        "type": "object",
        "required": [
          "markdown"
        ],
        "title": "ArtifactUpdateRequest"
      },
...
      "Body_upload_file_api_v1_threads__thread_id__files_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_upload_file_api_v1_threads__thread_id__files_post"
      },
...
      "DecisionRecordsResponse": {
        "properties": {
          "records": {
            "items": {
              "$ref": "#/components/schemas/DecisionRecord"
            },
            "type": "array",
            "title": "Records"
          }
        },
        "type": "object",
        "required": [
          "records"
        ],
        "title": "DecisionRecordsResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "HumanReviewAction": {
        "properties": {
          "suggestion_id": {
            "type": "string",
            "title": "Suggestion Id"
          },
          "action": {
            "$ref": "#/components/schemas/HumanReviewActionType"
          },
          "edited_suggestion": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Edited Suggestion"
          },
          "reviewer_id": {
            "type": "string",
            "title": "Reviewer Id",
            "default": "default-user"
          },
          "comment": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Comment"
          }
        },
        "type": "object",
        "required": [
          "suggestion_id",
          "action"
        ],
        "title": "HumanReviewAction"
      },
...
      "ReviewBatchResponse": {
        "properties": {
          "review_batch": {
            "$ref": "#/components/schemas/ReviewBatch"
          }
        },
        "type": "object",
        "required": [
          "review_batch"
        ],
        "title": "ReviewBatchResponse"
      },
...
      "ReviewSubmitRequest": {
        "properties": {
          "submitter_id": {
            "type": "string",
            "title": "Submitter Id",
            "default": "default-user"
          },
          "review_actions": {
            "items": {
              "$ref": "#/components/schemas/HumanReviewAction"
            },
            "type": "array",
            "title": "Review Actions"
          }
        },
        "type": "object",
        "required": [
          "review_actions"
        ],
        "title": "ReviewSubmitRequest"
      },
      "ReviewSubmitResponse": {
        "properties": {
          "submitted": {
            "type": "boolean",
            "title": "Submitted",
            "default": true
          },
          "review_batch_id": {
            "type": "string",
            "title": "Review Batch Id"
          }
        },
        "type": "object",
        "required": [
          "review_batch_id"
        ],
        "title": "ReviewSubmitResponse"
      },
...
      "ThreadFilesResponse": {
        "properties": {
          "files": {
            "items": {
              "$ref": "#/components/schemas/SourceDocument"
            },
            "type": "array",
            "title": "Files"
          }
        },
        "type": "object",
        "required": [
          "files"
        ],
        "title": "ThreadFilesResponse"
      },
...
      "UploadCategory": {
        "type": "string",
        "enum": [
          "context",
          "package"
        ],
        "title": "UploadCategory"
      },
      "UploadFileResponse": {
        "properties": {
          "uploaded": {
            "type": "boolean",
            "title": "Uploaded",
            "default": true
          },
          "filename": {
            "type": "string",
            "title": "Filename"
          },
          "category": {
            "$ref": "#/components/schemas/UploadCategory"
          }
        },
        "type": "object",
        "required": [
          "filename",
          "category"
        ],
        "title": "UploadFileResponse"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}